CIO-59 Organize utility functions and formatting
