+++ index.js
@@ -9,7 +9,7 @@ const updateNotifier = require('update-notifier');
 
 const git = require(path.resolve(__dirname, 'utils/git'));
 const dialogue = require(path.resolve(__dirname, 'utils/interface'));
-const { THEME_COLOR } = require(path.resolve(__dirname, 'utils/theme'));
+// const { THEME_COLOR } = require(path.resolve(__dirname, 'utils/theme'));
 
 // Checks for available update and returns an instance
 const pkg = require(path.resolve(__dirname, 'package.json'));
@@ -42,7 +42,7 @@ if (!process.argv.slice(2).length) {
   screen.key('r', () => {
     branchTable.clearItems();
 
-    git.fetchBranches().then(() => refreshTable());
+    git.doFetchBranches().then(() => refreshTable());
   });
 
   screen.append(branchTable);
@@ -54,9 +54,6 @@ if (!process.argv.slice(2).length) {
   process.on('SIGWINCH', () => {
     screen.emit('resize');
   });
-
-
-  // Handle key presses
 
   /**
    * @todo: Handle select -- differenciate spacebar or enter
@@ -103,15 +100,15 @@ if (!process.argv.slice(2).length) {
 
         if (answer === 'Yes') {
           await git
-            .checkoutBranch(gitBranch, gitRemote)
-            .then(git.createBranch(gitBranch))
+            .doCheckoutBranch(gitBranch, gitRemote)
+            .then(git.doCreateBranch(gitBranch))
             .then(screen.destroy());
         } else if (answer === 'No') {
-          await git.checkoutBranch(gitBranch, gitRemote).then(screen.destroy());
+          await git.doCheckoutBranch(gitBranch, gitRemote).then(screen.destroy());
         }
       });
     } else {
-      await git.checkoutBranch(gitBranch, gitRemote).then(screen.destroy());
+      await git.doCheckoutBranch(gitBranch, gitRemote).then(screen.destroy());
     }
   });
